requirements.txt
numpy
tbapy
mysql-connector-python
pytz
regex
